refactor(compact-index): promote CompactIndexClient to its own provider
The compact index (/info/) is a generic RubyGems protocol, not Artifactory-
specific: rubygems.org, a direct Contribsys/Gemstash/geminabox source, and
Artifactory all serve it. It was nested inside ArtifactoryClient only because
that was its first consumer. Lift it to StillActive::CompactIndexClient in its
own file, per the provider-per-file convention, so a second consumer (auditing a
direct compact-index source) composes it rather than reaching into Artifactory.

Behaviour-preserving move: the module is byte-identical, the caller resolves it
by constant lookup, and ArtifactoryClient keeps ownership of jfrog auth (the
client stays auth-agnostic, taking headers from its caller). Suite unchanged at
1268 examples.
